Enhanced patch to disable setDefaultDataSubId
This patch disables the function setDefaultDataSubId
ONLY for devices which have specifically SET,
or ADDED the flag "nodefaultdata" to the build.prop
variable "ro.telephony.ril.config".

In all other cases the outcome of the if statement
will return true and thus enable the function
setDefaultDataSubId.

This patch is needed for at least the gproj-common
 based devices, otherwise RIL doesn't function
(SIM Card is unable to be recognized).

A more permanent fix is being worked on,
but not to be expected very soon.
With this patch at least gproj devices can be updated,
so they can enjoy the latest security patches for android.

Change-Id: Ib124b35da51afd1cad9c3f38e42d44f48cc40ff8